+++ tgui/packages/tgui/interfaces/Newspaper.tsx
@@ -2,7 +2,7 @@ import { BooleanLike } from '../../common/react';
 import { useBackend } from '../backend';
 import { Box, Button, Divider, Image, Section } from '../components';
 import { Window } from '../layouts';
-import { processedText } from './Newscaster';
+import { processedText } from '../process';
 
 type Data = {
   current_page: number;
