+++ tgui/packages/tgui/process.ts
@@ -1,0 +1,17 @@
+import { marked } from 'marked';
+
+import { sanitizeText } from './sanitize';
+
+export const processedText = (value) => {
+  const textHtml = {
+    __html: sanitizeText(
+      marked(value, {
+        breaks: true,
+        smartypants: true,
+        smartLists: true,
+        baseUrl: 'thisshouldbreakhttp',
+      }),
+    ),
+  };
+  return textHtml;
+};
